# docs/swagger.yaml
definitions:
  oauth2.Token:
    properties:
      access_token:
        description: |-
          AccessToken is the token that authorizes and authenticates
          the requests.
        type: string
      expires_in:
        description: |-
          ExpiresIn is the OAuth2 wire format "expires_in" field,
          which specifies how many seconds later the token expires,
          relative to an unknown time base approximately around "now".
          It is the application's responsibility to populate
          `Expiry` from `ExpiresIn` when required.
        type: integer
      expiry:
        description: |-
          Expiry is the optional expiration time of the access token.

          If zero, [TokenSource] implementations will reuse the same
          token forever and RefreshToken or equivalent
          mechanisms for that TokenSource will not be used.
        type: string
      refresh_token:
        description: |-
          RefreshToken is a token that's used by the application
          (as opposed to the user) to refresh the access token
          if it expires.
        type: string
      token_type:
        description: |-
          TokenType is the type of token.
          The Type method returns either this or "Bearer", the default.
        type: string
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
  reqmodel.CommentCreateRequest:
    properties:
      movie_id:
        type: string
      text:
        type: string
    type: object
  reqmodel.CommentUpdateRequest:
    properties:
      text:
        type: string
    type: object
...
  reqmodel.MovieCommentListResponse:
    properties:
      movie_comment_list:
        items:
          $ref: '#/definitions/sqlc.GetMovieCommentListRow'
        type: array
      movie_id:
        type: string
    type: object
...
  reqmodel.UserCommentListResponse:
    properties:
      user_comment_list:
        items:
          $ref: '#/definitions/sqlc.GetUserCommentListRow'
        type: array
      user_id:
        type: string
    type: object
...
  sqlc.Comment:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      movie_id:
        type: string
      text:
        type: string
      user_id:
        type: string
    type: object
...
  sqlc.GetMovieCommentListRow:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      text:
        type: string
      user_id:
        type: string
    type: object
...
  sqlc.GetUserCommentListRow:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      movie_id:
        type: string
      text:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - multipart/form-data
      description: Get movie by id
      parameters:
      - description: Login
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth2.Token'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Auth
      tags:
      - auth
  /comment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Comment'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Create comments
      tags:
      - comment
  /comment/{comment_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Delete comment
      tags:
      - comment
      - admin
    get:
      consumes:
      - application/json
      description: Get comment by id
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Comment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get comment
      tags:
      - comment
    patch:
      consumes:
      - application/json
      parameters:
      - description: movie ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Comment update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.CommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Comment'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Update comments
      tags:
      - comment
...
  /movie/{movie_id}/comment:
    get:
      consumes:
      - application/json
      description: Get comment list for certain movie
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.MovieCommentListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie comments list
      tags:
      - comment
      - movie
...
  /user/{user_id}/comment:
    get:
      consumes:
      - application/json
      description: Get comment list for certain user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.UserCommentListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user comments list
      tags:
      - comment
      - user
...
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: http://localhost:8080/auth/login
    type: oauth2
swagger: "2.0"